% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_maps.R
\name{create_maps}
\alias{create_maps}
\title{create_maps}
\usage{
create_maps(
  data,
  zones = NULL,
  protocol,
  ices_data = readRDS(system.file("script_origin", "Data", "ICES_areas",
    "ICES_areas2.Rds", package = "inser"))
)
}
\arguments{
\item{data}{data.frame A dataframe as generated by prep_sel_data()}

\item{zones}{character The vector of zones used in the report. Default to NULL uses all the available area}

\item{protocol}{character Type of protocol. “twin” for experiment where both fishing gear are deployed in the same fishing operation (e.g. using twin trawls) ; “paired” for experiment where the two fishing gear are deployed in a set of ‘pairs’ : two different fishing operations, but in very similar conditions; “unpaired” for experiment where the gear are tested independently in two different samples of fishing operations.}

\item{ices_data}{sp object The spatial polygon data.frame of the study area}
}
\value{
ggplot A ggplot object of the map and zoom graphs
}
\description{
create_maps
}
\examples{
# Setup input OTT data
OTT_data_folder <- system.file("script_origin","Data","Example_OTT", package = "inser")

TR <- readr::read_delim(
  file = file.path(OTT_data_folder, "TR.csv"),
  delim = ";",
  escape_double = FALSE,
  locale = readr::locale(encoding = "WINDOWS-1252"),
  trim_ws = TRUE
)

HH<-read.table(
  file.path(OTT_data_folder, "HH.csv"),
  sep=";",
  header=TRUE,
  encoding = "WINDOWS-1252")#,colClasses = colClasses)

SL<-read.table(
  file.path(OTT_data_folder, "SL.csv"),
  sep=";",
  header=TRUE,
  encoding = "WINDOWS-1252")

HL<-read.table(
  file.path(OTT_data_folder, "HL.csv"),
  sep=";",
  header=TRUE,
  encoding = "WINDOWS-1252")

colClasses<-rep(NA,ncol(HH))
colClasses[which(names(HH)=="statistical_rectangle")]<-"character"

HH<-read.table(
  file.path(OTT_data_folder, "HH.csv"),
  sep=";",
  header=TRUE,
  colClasses = colClasses,
  encoding = "WINDOWS-1252")

# create TAB output
TAB <- prep_sel_data(data=list(TR,HH,SL,HL))

# create maps
output <- create_maps(
  zones = c("8.a","8.b","7.d","7.e","7.h"),
  protocol = "twin",
  data = TAB
)
}
