% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_translate_dict.R
\name{create_translate_dict}
\alias{create_translate_dict}
\title{create_translate_dict}
\usage{
create_translate_dict(
  path = system.file("template", "translation.csv", package = "inser"),
  language = c("EN", "FR"),
  encoding = "UTF-8",
  sep = ";"
)
}
\arguments{
\item{path}{character. The path to the csv translation file}

\item{language}{character. The chosen language for the text, must be "FR" or "EN" (default)}

\item{encoding}{character. The encoding of the csv file, default to "UTF-8"}

\item{sep}{character. The field separator used in the CSV file}
}
\value{
list. A list of translated text, with names referring to their location in the sheet
}
\description{
Create a list with all the text elements of the selectivity sheet in the selected language
}
\examples{
# create the EN translator
lg <- create_translate_dict(language = "EN")

# call main title by its ID
cat(lg[["0_main_title"]])

# create the FR translator
lg <- create_translate_dict(language = "FR")

# call main title by its ID
cat(lg[["0_main_title"]])
}
