% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_test.R
\name{f_test}
\alias{f_test}
\title{f_test
Computes the paired tests on difference in weight per catch category}
\usage{
f_test(tab_diff, sp, cat)
}
\arguments{
\item{tab_diff}{data.frame. The dataframe containing information on weight per catch category}

\item{sp}{character. The species to use in test}

\item{cat}{character. The category to use in test}
}
\value{
data.frame. The summary table of the statistical tests
}
\description{
f_test
Computes the paired tests on difference in weight per catch category
}
\examples{
# Create tmp folder
output_dir <- tempfile(pattern = "inser")
dir.create(output_dir)

# Setup input OTT data
OTT_data_folder <- system.file("script_origin","Data","Example_OTT", package = "inser")

TR <- readr::read_delim(
  file = file.path(OTT_data_folder, "TR.csv"),
  delim = ";",
  escape_double = FALSE,
  locale = readr::locale(encoding = "WINDOWS-1252"),
  trim_ws = TRUE
)

HH<-read.table(
  file.path(OTT_data_folder, "HH.csv"),
  sep=";",
  header=TRUE,
  encoding = "WINDOWS-1252")#,colClasses = colClasses)

SL<-read.table(
  file.path(OTT_data_folder, "SL.csv"),
  sep=";",
  header=TRUE,
  encoding = "WINDOWS-1252")

HL<-read.table(
  file.path(OTT_data_folder, "HL.csv"),
  sep=";",
  header=TRUE,
  encoding = "WINDOWS-1252")

colClasses<-rep(NA,ncol(HH))
colClasses[which(names(HH)=="statistical_rectangle")]<-"character"

HH<-read.table(
  file.path(OTT_data_folder, "HH.csv"),
  sep=";",
  header=TRUE,
  colClasses = colClasses,
  encoding = "WINDOWS-1252")

# create selectivity data object
data <- prep_sel_data(data=list(TR,HH,SL,HL))

# extract weight data per species
weight_species <- data \%>\%
  dplyr::group_by(
    project,
    vessel_identifier,
    trip_code,
    station_number,
    gear_label,
    catch_category,
    species
  ) \%>\%
  dplyr::summarize(weight = sum(weight) * 10 ^ (-3)) \%>\%
  as.data.frame()

weight_species <- as.data.frame(tidyr::complete(
  weight_species,
  tidyr::nesting(project, vessel_identifier, trip_code, station_number),
  gear_label,
  catch_category,
  species,
  fill = list(weight = 0)
))

tab_diff <- weight_species \%>\% dplyr::group_by(project,
                                               vessel_identifier,
                                               trip_code,
                                               station_number,
                                               catch_category,
                                               species) \%>\%
  dplyr::summarize(
    diff_weight = weight[gear_label == "TEST"] - weight[gear_label == "STD"],
    weight_STD = weight[gear_label == "STD"],
    weight_TEST = weight[gear_label == "TEST"]
  )

# run f_test
f_test(tab_diff = tab_diff,
       sp = "Solea solea",
       cat = "LAN")

# Clear tmp folder
unlink(output_dir, recursive = TRUE)
}
