% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_test_unpaired.R
\name{f_test_unpaired}
\alias{f_test_unpaired}
\title{f_test_unpaired}
\usage{
f_test_unpaired(weight_species, sp, cat)
}
\arguments{
\item{weight_species}{data.frame. A dataframe containing the weights of catch per category for TEST and STD device}

\item{sp}{character. The species to use in test}

\item{cat}{character. The category to use in test}
}
\value{
data.frame. The summary table of the statistical tests
}
\description{
Computes the unpaired tests on difference in weight per catch category
}
\examples{
# Setup OTB input data
OTB_data_folder <-
  system.file("script_origin", "Data", "Example_OTB_alternate", package = "inser")

TR <- readr::read_delim(
  file = file.path(OTB_data_folder, "TR.csv"),
  delim = ";",
  escape_double = FALSE,
  locale = readr::locale(encoding = "WINDOWS-1252"),
  trim_ws = TRUE
)
HH <-
  read.table(
    file.path(OTB_data_folder, "HH.csv"),
    sep = ";",
    header = TRUE,
    encoding = "WINDOWS-1252"
  )#,colClasses = colClasses)
SL <-
  read.table(
    file.path(OTB_data_folder, "SL.csv"),
    sep = ";",
    header = TRUE,
    encoding = "WINDOWS-1252"
  )
HL <-
  read.table(
    file.path(OTB_data_folder, "HL.csv"),
    sep = ";",
    header = TRUE,
    encoding = "WINDOWS-1252"
  )

colClasses <- rep(NA, ncol(HH))
colClasses[which(names(HH) == "statistical_rectangle")] <-
  "character"

HH <-
  read.table(
    file.path(OTB_data_folder, "HH.csv"),
    sep = ";",
    header = TRUE,
    colClasses = colClasses,
    encoding = "WINDOWS-1252"
  )

# create TAB output
data <- prep_sel_data(data = list(TR, HH, SL, HL))

# weight for each species
weight_species <-
  data \%>\% dplyr::group_by(
    project,
    vessel_identifier,
    trip_code,
    station_number,
    gear_label,
    catch_category,
    species
  ) \%>\%
  dplyr::summarize(weight = sum(weight) * 10 ^ (-3)) \%>\% as.data.frame()

weight_species <-
  as.data.frame(
    tidyr::complete(
      weight_species,
      tidyr::nesting(project, vessel_identifier, trip_code, station_number),
      gear_label,
      catch_category,
      species,
      fill = list(weight = 0)
    )
  )

# run f_test_unpaired
f_test_unpaired(weight_species = weight_species,
                sp = "Solea solea",
                cat = "LAN")

}
