% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inser-package.R
\docType{package}
\name{inser-package}
\alias{inser}
\alias{inser-package}
\title{inser: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Marie Morfin \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

Authors:
\itemize{
  \item Swann Floc'hlay \email{swann@thinkr.fr} (\href{https://orcid.org/0000-0003-1477-830X}{ORCID})
  \item Antoine Languillaume \email{antoine@thinkr.fr} (\href{https://orcid.org/0000-0001-9843-5632}{ORCID})
}

}
\keyword{internal}
