% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_selectivity_sheet.R
\name{create_selectivity_sheet}
\alias{create_selectivity_sheet}
\title{create_selectivity_sheet Title}
\usage{
create_selectivity_sheet(
  data,
  protocol,
  output_dir = NULL,
  output_file = NULL,
  min_length = NULL,
  language = "EN",
  zones = NULL
)
}
\arguments{
\item{data}{data.frame A dataframe as generated by prep_sel_data()}

\item{protocol}{character Type of protocol. “twin” for experiment where both fishing gear are deployed in the same fishing operation (e.g. using twin trawls) ; “paired” for experiment where the two fishing gear are deployed in a set of ‘pairs’ : two different fishing operations, but in very similar conditions; “unpaired” for experiment where the gear are tested independently in two different samples of fishing operations.}

\item{output_dir}{character Directory name where the sheet should be written}

\item{output_file}{character Name of the file for the sheet}

\item{min_length}{numeric Optional. For species analyzed in length, the minimum targeted length for selectivity (e.g., MCRS) for graphics and statistics. A data.frame with a column ‘species’ and a column ‘min_length’ for the corresponding minimum length.}

\item{language}{character English by default (“EN”). Can be passed in French (“FR”).}

\item{zones}{character The vector of zones used in the report. Default to NULL uses all the available area}
}
\value{
None A Word document is generated in the output_dir as well as all the figures in png.
}
\description{
Génère une fiche synthétique décrivant les données et contenant
un ensemble d’indicateurs de sélectivité.
}
\examples{
# Create tmp folder
output_dir <- tempfile(pattern = "inser")
dir.create(output_dir)

# Setup input OTT data
OTT_data_folder <- system.file("script_origin","Data","Example_OTT", package = "inser")

TR <- readr::read_delim(
  file = file.path(OTT_data_folder, "TR.csv"),
  delim = ";",
  escape_double = FALSE,
  locale = readr::locale(encoding = "WINDOWS-1252"),
  trim_ws = TRUE
)

HH<-read.table(
  file.path(OTT_data_folder, "HH.csv"),
  sep=";",
  header=TRUE,
  encoding = "WINDOWS-1252")#,colClasses = colClasses)

SL<-read.table(
  file.path(OTT_data_folder, "SL.csv"),
  sep=";",
  header=TRUE,
  encoding = "WINDOWS-1252")

HL<-read.table(
  file.path(OTT_data_folder, "HL.csv"),
  sep=";",
  header=TRUE,
  encoding = "WINDOWS-1252")

colClasses<-rep(NA,ncol(HH))
colClasses[which(names(HH)=="statistical_rectangle")]<-"character"

HH<-read.table(
  file.path(OTT_data_folder, "HH.csv"),
  sep=";",
  header=TRUE,
  colClasses = colClasses,
  encoding = "WINDOWS-1252")

# create TAB output
TAB <- prep_sel_data(data=list(TR,HH,SL,HL))

# Setup zone and min_length params
min_length <- data.frame(species=unique(TAB$species),min_length=c(27,24,20,NA,27,NA))
zones <- c("8.a","8.b","7.d","7.e","7.h")

# run function
create_selectivity_sheet(data = TAB,
                         output_dir = output_dir,
                         output_file = "fiche_InseR_twin",
                         protocol = "twin",
                         language = "FR",
                         zones=zones,
                         min_length=min_length)

# Clear tmp folder
unlink(output_dir, recursive = TRUE)
}
