% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inser-package.R
\docType{package}
\name{inser-package}
\alias{inser}
\alias{inser-package}
\title{inser: Generate Fishing Gear Selectivity Indicator Sheets}
\description{
Generate a synthetic sheet of fishing gear selectivity indicators (statistics, graphics and tests) from catch comparison data collected at sea. These data can be obtained from three types of experimental protocols: 'twin', 'single paired' or 'independent' sampling, for which a test gear is compared to a standard gear.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mariemorfin/inser/}
}

}
\author{
\strong{Maintainer}: Marie Morfin \email{first.last@example.com} (\href{https://orcid.org/0000-0002-1637-5910}{ORCID})

Authors:
\itemize{
  \item Swann Floc'hlay \email{swann@thinkr.fr} (\href{https://orcid.org/0000-0003-1477-830X}{ORCID})
  \item Antoine Languillaume \email{antoine@thinkr.fr} (\href{https://orcid.org/0000-0001-9843-5632}{ORCID})
}

}
\keyword{internal}
