% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_sel_data.R
\name{prep_sel_data}
\alias{prep_sel_data}
\title{prep_sel_data}
\usage{
prep_sel_data(
  data,
  filters = list(project = "all", vessel = "all", trip = "all", station = "all",
    species_LAN = "all", species_DIS = "all", species_length = "all")
)
}
\arguments{
\item{data}{list A list of four data.frames (TR,HH,SL,HL) based on the ICES RDB data exchange format corresponding to each level of the dataset. EU/ICES have defined common format and processing tools, for fisheries statistics.}

\item{filters}{list A list of filters to apply on the dataset. The optional filters can apply directly on the fields ‘project’, ‘vessel_identifier’, ‘selective_device’, ‘trip_code’, ‘station_number’. For the fields ‘species_LAN’, ‘species_DIS’, and ‘species_length’, the filter apply on the landings (LAN), discards (DIS) and on the measured species, respectively.}
}
\value{
data.frame A data.frame object corresponding to the join of the four tables.
}
\description{
Transforme et prépare base de donnée pour le calcul des indicateurs dans la fonction create_selectivity_sheet(). Les données d’entrée sont issues d’observations de captures en mer dans le cadre d’une expérimentation pour comparer la sélectivité d’un engin de pêche test par rapport à un engin standard (de référence).  Typiquement les données issues de ce type d’expérimentation sont saisies sous forme d’une base de données composées de 4 tables : TR correspond aux informations relatives au duo marée x engin ; HH correspond au duo opération de pêche x engin ; SL correspond aux informations de poids une fois le tri de la capture réalisé : par partie de la capture (rejet/débarquement) et espèce (obligatoires), et éventuellement également par catégorie commerciale, sexe. HL correspond aux effectifs par classe de taille. Toutes les espèces ne sont pas forcément pesées ni mesurées. S'assure de la cohérence entre les 4 tables. Filtre la base sur les bateaux/marées/stations/espèces sur lesquelles réaliser la fiche. Calcul des effectifs élevés à partir des fractions échantillonnées si les informations manquent.
}
\details{
If the field ‘elevated_number_at_length’ is not provided in HL or some values are missing, the function calculate them using ‘number_at_length’ and the ratio between the subsample_weight and the weight (table SL).
}
\examples{
### Example for protocol 'twin'

OTT_data_folder <- system.file("script_origin", "Data", "Example_OTT",
                               package = "inser"
)

TR <- readr::read_delim(
  file = file.path(OTT_data_folder, "TR.csv"),
  delim = ";",
  escape_double = FALSE,
  locale = readr::locale(encoding = "WINDOWS-1252"),
  trim_ws = TRUE
)

HH <- read.table(
  file.path(OTT_data_folder, "HH.csv"),
  sep = ";",
  header = TRUE,
  encoding = "WINDOWS-1252"
)#,colClasses = colClasses)

SL <- read.table(
  file.path(OTT_data_folder, "SL.csv"),
  sep = ";",
  header = TRUE,
  encoding = "WINDOWS-1252"
)

HL <- read.table(
  file.path(OTT_data_folder, "HL.csv"),
  sep = ";",
  header = TRUE,
  encoding = "WINDOWS-1252"
)

colClasses <- rep(NA, ncol(HH))
colClasses[which(names(HH) == "statistical_rectangle")] <- "character"

HH <- read.table(
  file.path(OTT_data_folder, "HH.csv"),
  sep = ";",
  header = TRUE,
  colClasses = colClasses,
  encoding = "WINDOWS-1252"
)

TAB <- prep_sel_data(data = list(TR, HH, SL, HL))
}
